Refactor Bookmarks Sync into Dedicated Service
- Extract bookmark synchronization logic from BookmarkManager to a new bookmarks service
- Add syncBookmarks function in src/services/bookmarks.ts
- Update AuthContext and BookmarkManager to use new sync service
- Simplify sync handling and improve code modularity
- Maintain existing sync functionality with cloud and local bookmarks
